requirements.txt
openpyxl~=3.1.2
python-taiga
python-dotenv~=1.0.1
openai

Django~=5.0.3
environ~=1.0
transliterate
bcrypt~=4.1.2

djangorestframework~=3.15.2